Adopt size_t in more places and remove compiler warnings
- change Data::size_type from uint32_t to size_t to more inline with C++ API standard - also helps remove 64-bit compilation warnings
- ares.h - fix include order issue that was causing redefinition errors or NO_ERROR on windows builds

+++ apps/sipdial/DialInstance.cpp
@@ -167,8 +167,8 @@ void DialInstance::sendInvite()
 Data DialInstance::processNumber(const Data& verboseNumber)
 {
    Data num = Data("");
-   int len = verboseNumber.size();
-   for(int i = 0; i < len; i++)
+   size_t len = verboseNumber.size();
+   for(size_t i = 0; i < len; i++)
    {
       char c = verboseNumber[i];
       switch(c)
