+++ resip/recon/MOHParkServer/HttpConnection.cxx
@@ -337,7 +337,7 @@ HttpConnection::processSomeWrites()
    //DebugLog (<< "Writing " << mTxBuffer );
 
 #if defined(WIN32)
-   int bytesWritten = ::send( mSock, mTxBuffer.data(), mTxBuffer.size(), 0);
+   int bytesWritten = ::send( mSock, mTxBuffer.data(), (int)mTxBuffer.size(), 0);
 #else
    int bytesWritten = ::write(mSock, mTxBuffer.data(), mTxBuffer.size() );
 #endif
