+++ resip/stack/MultipartMixedContents.cxx
@@ -176,7 +176,7 @@ MultipartMixedContents::parse(ParseBuffer& pb)
    boundaryNoCRLF += boundaryToken;
 
    pb.skipToChars(boundaryNoCRLF);
-   pb.skipN((int)boundaryNoCRLF.size());
+   pb.skipN(boundaryNoCRLF.size());
    pb.assertNotEof();
 
    do
@@ -242,7 +242,7 @@ MultipartMixedContents::parse(ParseBuffer& pb)
       ParseBuffer headersPb(headerStart, bodyStart-4-headerStart);
       mContents.back()->preParseHeaders(headersPb);
 
-      pb.skipN((int)boundary.size());
+      pb.skipN(boundary.size());
 
       const char* loc = pb.position();
       pb.skipChar();
